+++ tests/testthat/test-to_null.R
@@ -1,8 +1,8 @@
-test_that(".to_null() works on the happy path", {
+test_that(".to_null() works on the happy path (#129)", {
   expect_null(.to_null(NULL))
 })
 
-test_that(".to_null() errors when NULL isn't allowed", {
+test_that(".to_null() errors when NULL isn't allowed (#129)", {
   given <- NULL
   expect_error(
     .to_null(given, allow_null = FALSE),
@@ -18,14 +18,14 @@ test_that(".to_null() errors when NULL isn't allowed", {
   )
 })
 
-test_that(".to_null() coerces anything to NULL", {
+test_that(".to_null() coerces anything to NULL (#129)", {
   expect_null(.to_null(1L))
   expect_null(.to_null(mean))
   expect_null(.to_null(TRUE))
   expect_null(.to_null(letters))
 })
 
-test_that(".to_null() errors for bad allow_null", {
+test_that(".to_null() errors for bad allow_null (#129)", {
   expect_error(
     .to_null(NULL, allow_null = NULL),
     class = .compile_dash("stbl", "error", "bad_null")
@@ -49,7 +49,7 @@ test_that(".to_null() errors for bad allow_null", {
   )
 })
 
-test_that(".to_null() errors informatively for missing value", {
+test_that(".to_null() errors informatively for missing value (#129)", {
   expect_error(
     .to_null(),
     class = .compile_dash("stbl", "error", "must")